Review & refactor duplications rules
