+++ tests/linter/rules/duplications/both_tests_and_tasks/expected_output.txt
@@ -1,2 +1,2 @@
-test.robot:14:1 [E] DUP10 Both Task(s) and Test Case(s) section headers defined in file
-test2.robot:3:1 [E] DUP10 Both Task(s) and Test Case(s) section headers defined in file
+test.robot:14:1 [E] DUP09 Both Task(s) and Test Case(s) section headers defined in file
+test2.robot:3:1 [E] DUP09 Both Task(s) and Test Case(s) section headers defined in file
