--- tests/linter/rules/duplications/duplicated_argument_name/expected_output.txt
@@ -1,5 +1,0 @@
-test.robot:3:30 [E] DUP11 Argument name '${var}' is already used
-test.robot:4:12 [E] DUP11 Argument name '${var}' is already used
-test.robot:4:22 [E] DUP11 Argument name '${Var}' is already used
-test.robot:4:32 [E] DUP11 Argument name '${var}' is already used
-test.robot:15:42 [E] DUP11 Argument name '${PARAM}' is already used
