+++ tests/linter/rules/duplications/duplicated_variable/expected_output.txt
@@ -1,2 +1,2 @@
-test.robot:8:1 [E] DUP03 Multiple variables with name '${V AR}' in Variables section (first occurrence in line 6). Note that Robot Framework is case-insensitive
-test.robot:9:1 [E] DUP03 Multiple variables with name '${other_var}' in Variables section (first occurrence in line 7). Note that Robot Framework is case-insensitive
+test.robot:8:1 [E] DUP03 Multiple variables with name '${V AR}' in Variables section (first occurrence in line 6)
+test.robot:9:1 [E] DUP03 Multiple variables with name '${other_var}' in Variables section (first occurrence in line 7)
