--- tests/linter/rules/duplications/section_out_of_order/expected_output.txt
@@ -1,2 +1,0 @@
-test.robot:20:1 [W] DUP09 '*** Settings ***' section header is defined in wrong order: Settings > Variables > Test Cases / Tasks > Keywords
-test.robot:23:1 [W] DUP09 '*** Variables ***' section header is defined in wrong order: Settings > Variables > Test Cases / Tasks > Keywords
