+++ tests/linter/rules/order/section_out_of_order/custom_order.robot
