+++ tests/linter/rules/variables/duplicated_assigned_var_name/test.robot
